requirements.txt
django
channels["daphne"]
djangorestframework
markdown
django-filter